# mypaintings/target/classes/api-spec.yaml
openapi: 3.0.0
info:
  title: MyP@intings
  version: 1.0.0
  description: RESTful API dla aplikacji myp@intings
...
paths:
...
  /gosc/reserve:
    post:
      tags:
        - Gosc
      summary: Zarezerwuj obraz (Gosc)
      description: Endpoint pozwala gościowi zarezerwować obraz.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request_gosc_reserve'
      responses:
        '200':
          description: Rezerwacja zakończona sukcesem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_gosc_reserve'
        '400':
          description: Nieprawidłowe dane wejściowe
        '401':
          description: Brak autoryzacji
        '404':
          description: Obraz nie znaleziony
  /gosc/zamowienia:
    get:
      tags:
        - Gosc
      summary: Pobierz listę zamówień gościa
      description: Endpoint zwraca listę zamówień danego gościa. Wymaga tokenu JWT.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_gosc_zamowienia'
        '401':
          description: Brak autoryzacji
        '403':
          description: Brak dostępu
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    response_gosc_reserve:
      type: object
      properties:
        message:
          type: string
          example: "Obraz został zarezerwowany."
    response_gosc_zamowienia:
      type: object
      properties:
        zamowienia:
          type: array
          items:
            type: object
            properties:
              bought_on:
                type: string
                format: date
                example: "2024-12-11"
              price:
                type: string
                example: "$650"
              hotel:
                type: string
                example: "@Hotel Lord"
              status:
                type: string
                example: "Waiting for payment (reception)"
              image_url:
                type: string
                example: "https://example.com/images/image.png"
...
    request_gosc_reserve:
      type: object
      properties:
        image_id:
          type: integer
          example: 123
        buyer_id:
          type: integer
          nullable: true
          example: null
      required:
        - image_id
        - buyer_id
...